Add wav conversion, fix mp3 copy

+++ main.rb
@@ -8,8 +8,11 @@ require 'gli'
 require_relative './lib/config'
 require_relative './lib/beats'
 require_relative './lib/beats/destination_track_file'
+require_relative './lib/beats/flac_conversion'
 require_relative './lib/beats/library'
 require_relative './lib/beats/track_file'
+require_relative './lib/beats/vinyl_conversion'
+require_relative './lib/beats/wav_conversion'
 
 class App
   extend GLI::App
@@ -33,6 +36,7 @@ class App
       app.process_flacs
       app.process_aiffs
       app.process_mp3s
+      app.process_wavs
     end
   end
 
@@ -45,6 +49,7 @@ class App
     dest = File.join(TRACKS_PATH, filename)
     source_file = Beats::TrackFile.read library: library, path: path
     dest_file = Beats::DestinationTrackFile.new album: source_file.album, track: source_file.track, ext: source_file.ext
+    return if dest_file.exist?
     dest_file.ensure_dest_path!
     puts "#{source_file.path} => #{dest_file.path}"
     FileUtils.cp source_file.path, dest_file.path
@@ -62,6 +67,15 @@ class App
   def process_flacs
     Beats.each_flac do |path|
       converter = Beats::FlacConversion.from_file library: library, path: path
+      if converter.process!
+        puts "#{converter.dest_file.album.title} - #{converter.dest_file.track.label}..."
+      end
+    end
+  end
+
+  def process_wavs
+    Beats.each_wav do |path|
+      converter = Beats::WavConversion.from_file library: library, path: path
       if converter.process!
         puts "#{converter.dest_file.album.title} - #{converter.dest_file.track.label}..."
       end
